+++ CHANGELOG.md
@@ -4,7 +4,8 @@ Changelog
 0.2
 ---
 
-  * ...
+  * Added Ppx_deriving.expand_path.
+  * Change Ppx_deriving.deriver signature to pass ~options and ~path as labeled arguments.
 
 0.1
 ---